Move storage utils into app
