+++ TrinityBackendFastAPI/app/features/data_upload_validate/app/routes.py
@@ -102,19 +102,25 @@ async def health_check():
 from minio import Minio
 from minio.error import S3Error
 from app.features.feature_overview.deps import redis_client
-from contexts.DataStorageRetrieval import (
+from app.utils.db import (
     fetch_client_app_project,
     record_arrow_dataset,
     rename_arrow_dataset,
     delete_arrow_dataset,
     arrow_dataset_exists,
-    upload_dataframe,
+)
+from app.utils.arrow_client import upload_dataframe
+from app.utils.flight_registry import (
     set_ticket,
     get_ticket_by_key,
     get_latest_ticket_for_basename,
     get_original_csv,
     rename_arrow_object,
     remove_arrow_object,
+    get_flight_path_for_csv,
+    get_arrow_for_flight_path,
+)
+from app.utils.minio_utils import (
     ensure_minio_bucket,
     save_arrow_table,
     upload_to_minio,
